+++ tests/test_level4.py
@@ -1,4 +1,3 @@
-
 import unittest
 
 import DiceStatistics as tts
